+++ tests/testthat/test-across.R
@@ -970,6 +970,39 @@ test_that("selects and combines columns", {
   expect_equal(out$z, list(1:4))
 })
 
+test_that("can't rename during selection (#6522)", {
+  df <- tibble(x = 1)
+
+  expect_snapshot(error = TRUE, {
+    mutate(df, z = c_across(c(y = x)))
+  })
+})
+
+test_that("can't explicitly select grouping columns (#6522)", {
+  # Related to removing the mask layer from the quosure environments
+  df <- tibble(g = 1, x = 2)
+  gdf <- group_by(df, g)
+
+  expect_snapshot(error = TRUE, {
+    mutate(gdf, y = c_across(g))
+  })
+})
+
+test_that("`all_of()` is evaluated in the correct environment (#6522)", {
+  # Related to removing the mask layer from the quosure environments
+  df <- tibble(x = 1, y = 2)
+
+  expect_snapshot(error = TRUE, {
+    mutate(df, z = c_across(all_of(y)))
+  })
+
+  y <- "x"
+  expect <- df[["x"]]
+
+  out <- mutate(df, z = c_across(all_of(y)))
+  expect_identical(out$z, expect)
+})
+
 # cols deprecation --------------------------------------------------------
 
 test_that("across() applies old `.cols = everything()` default with a warning", {