+++ NEWS.md
@@ -1,5 +1,11 @@
 # dplyr (development version)
 
+* `c_across()` now evaluates `all_of()` correctly and no longer allows you to
+  accidentally select grouping variables (#6522).
+
+* `c_across()` now throws a more informative error if you try to rename during
+  column selection (#6522).
+
 * `n_distinct()` now errors if you don't give it any input (#6535).
 
 * `group_walk()` gains an explict `.keep` argument (#6530).